+++ inc/ti/mod/github.h
@@ -5,19 +5,16 @@
 #define TI_MOD_GITHUB_H_
 
 #include <ex.h>
-
-typedef struct ti_mod_github_s ti_mod_github_t;
+#include <stdbool.h>
+#include <curl/curl.h>
+#include <ti/module.t.h>
+#include <ti/mod/github.t.h>
 
 _Bool ti_mod_github_test(const char * s, size_t n);
-ti_mod_github_t * ti_mod_github_create(const char * s, size_t n);
-
-struct ti_mod_github_s
-{
-    char * owner;
-    char * repo;
-    char * token;  /* NULL when public, not visible */
-    char * ref;  /* tag, branch etc. */
-};
+ti_mod_github_t * ti_mod_github_create(const char * s, size_t n, ex_t * e);
+void ti_mod_github_destroy(ti_mod_github_t * gh);
+void ti_mod_github_install(ti_module_t * module);
+const char * ti_mod_github_code_str(ti_mod_github_t * gh);
 
 
 #endif  /* TI_MOD_GITHUB_H_ */