+++ main.c
@@ -23,6 +23,7 @@
 #include <time.h>
 #include <util/fx.h>
 #include <util/logger.h>
+#include <curl/curl.h>
 
 
 static void main__init_deploy(void)
@@ -59,6 +60,9 @@ int main(int argc, char * argv[])
 
     /* set local to LC_ALL and C to force a period over comma for float */
     (void) setlocale(LC_ALL, "C");
+
+    /* initialize global curl */
+    curl_global_init(CURL_GLOBAL_ALL);
 
     /* initialize random */
     srand(time(NULL));
@@ -258,6 +262,10 @@ stop:
     }
 
     ti_destroy();
+
+    /* cleanup global curl */
+    curl_global_cleanup();
+
     if (rc)
         log_error("exit with error code %d", rc);
     else